+++ .github/workflows/pr.yaml
@@ -22,7 +22,7 @@ jobs:
           DIFF_FILE=$(realpath diff.txt)
           cd current
           yarn
-          ./src/diff.ts ./src/endpoints.d.ts ../base/src/endpoints.d.ts > $DIFF_FILE
+          ./src/diff.ts ../base/api/index.ts ./api/index.ts > $DIFF_FILE
           echo "DIFF_FILE=$DIFF_FILE" >> $GITHUB_ENV
 
       - uses: actions/github-script@v5
