+++ api/index.ts
@@ -1,0 +1,4 @@
+import {Endpoints as TerminalEndpoints, Resources as TerminalResources} from "./terminal";
+
+export type Endpoints = TerminalEndpoints;
+export type Resources = TerminalResources;
